Fix pagination info object registrations
