+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Fixed
+
+- Add `PaginatorInfo`, `SimplePaginatorInfo`, `PageInfo` only as soon as they're required
+
 ## v6.11.0
 
 ### Added
